moved shapeLog to LogsViewer component
