+++ src/components/Transaction/LogsViewer.vue
@@ -17,7 +17,7 @@
             Human-readable logs
         </div>
         <div class="col-12">
-            <logs-table v-if="human_readable" :logs="logs" :contract="contract"/>
+            <logs-table v-if="human_readable" :logs="shapedLogs" :contract="contract"/>
             <json-viewer
                 v-else
                 :value="logs"
@@ -32,6 +32,7 @@
 <script>
 import JsonViewer from 'vue-json-viewer'
 import LogsTable from 'components/Transaction/LogsTable'
+import { TRANSFER_FUNCTION_SIGNATURES } from 'src/lib/functionSignatures';
 
 export default {
     name: 'LogsViewer',
@@ -49,11 +50,48 @@ export default {
             required: true,
         },
     },
-    mounted(){
-        console.log(this.logs);
+
+    created() {
+        let contracts = {};
+        this.logs.forEach(async (log) => {
+            let shapedLog = log;
+            if(!this.contract || log.address !== this.contract.address){
+                let log_contract;
+                if (Object.prototype.hasOwnProperty.call(contracts, log.address)){
+                    log_contract = contracts[log.address]
+                } else {
+                    const metadata = await this.$contractManager.checkBucket(log.address);
+                    if (metadata) {
+                        log_contract = await this.$contractManager.getVerifiedContract(log.address.toLowerCase(), metadata);
+                    } else {
+                        log_contract = await this.$contractManager.getContract(log.address.toLowerCase());
+                    }
+                    contracts[log.address] = log_contract;
+                }
+                if(log_contract){
+                    let logs = await log_contract.parseLogs([log]);
+                    shapedLog = logs[0];
+                    shapedLog.address = log.address;
+                    shapedLog.function_signature = shapedLog.topic ? shapedLog.topic.substr(0, 10) : shapedLog.topics[0].substr(0, 10);
+                    shapedLog.name = shapedLog.signature;
+                }
+            }
+
+            if (TRANSFER_FUNCTION_SIGNATURES.includes(shapedLog.function_signature)) {
+                shapedLog.isTransfer = true;
+                try {
+                    shapedLog.token = await this.$contractManager.getTokenData(log.address, 'erc20');
+                } catch (e) {
+                    console.error(`Failed to retrieve contract with address ${log.address}: ${e.message}`);
+                }
+            }
+
+            this.shapedLogs.push(shapedLog)
+        });
     },
     data: () => ({
         human_readable: true,
+        shapedLogs: [],
     }),
 }
 </script>
